+++ private/language/dutch.php
@@ -2059,7 +2059,8 @@ $LANG_featurecodes = array(
 
 $LANG_frontpagecodes = array(
     0 => 'Alleen in topic laten zien',
-    1 => 'Laat op voorpagina zien'
+    1 => 'Laat op voorpagina zien',
+    2 => 'Show on Front Page Until'
 );
 
 $LANG_postmodes = array(