+++ private/language/finnish_utf-8.php
@@ -2056,7 +2056,8 @@ $LANG_featurecodes = array(
 
 $LANG_frontpagecodes = array(
     0 => 'N&auml;yt&auml; Vain Aiheessa',
-    1 => 'N&auml;yt&auml; Etusivulla'
+    1 => 'N&auml;yt&auml; Etusivulla',
+    2 => 'Show on Front Page Until'
 );
 
 $LANG_postmodes = array(