+++ private/language/french_canada_utf-8.php
@@ -2057,7 +2057,8 @@ $LANG_featurecodes = array(
 
 $LANG_frontpagecodes = array(
     0 => 'Afficher Uniquement en Sujet',
-    1 => 'Afficher sur Front Page'
+    1 => 'Afficher sur Front Page',
+    2 => 'Show on Front Page Until'
 );
 
 $LANG_postmodes = array(