+++ private/language/turkish.php
@@ -2053,7 +2053,8 @@ $LANG_featurecodes = array(
 
 $LANG_frontpagecodes = array(
     0 => 'Sadece Konularda Göster',
-    1 => 'Anasayfada Göster'
+    1 => 'Anasayfada Göster',
+    2 => 'Show on Front Page Until'
 );
 
 $LANG_postmodes = array(