+++ public_html/admin/vercheck.php
@@ -44,7 +44,7 @@ if (!SEC_inGroup ('Root')) {
 
 function _displayVersionData()
 {
-    global $_CONF, $_USER, $LANG01, $LANG_UPGRADE, $LANG01, $LANG_FILECHECK, $LANG_ADMIN, $_PLUGIN_INFO;
+    global $_CONF, $_USER, $LANG01, $LANG_UPGRADE, $LANG_ENVCHK, $LANG01, $LANG_FILECHECK, $LANG_ADMIN, $_PLUGIN_INFO;
 
     $retval = '';
     $upToDate = 0;
@@ -116,7 +116,7 @@ function _displayVersionData()
         $T->set_var(array(
             'phpendoflife' => true,
             'phpeol_statusmsg' => $LANG_UPGRADE['phpeol'],
-            'phpeol_statustext' => $LANG01['phpendoflife'],
+            'phpeol_statustext' => $LANG_ENVCHK['phpendoflife'],
         ));
     }